requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# ML/NLP libraries
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
bertopic==0.15.0
umap-learn==0.5.4
hdbscan==0.8.33
scikit-learn==1.3.2

# Summarization
bert-score==0.3.13
rouge-score==0.1.2

# Data processing
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
spacy==3.7.2

# Database/Storage
chromadb==0.4.18

# Plotting
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
pydantic==2.5.0
python-multipart==0.0.6